requirement.txt
langchain
langgraph
langchain-core
langchain-community
ipykernel
python-dotenv
langchain-groq
langgraph-cli[inmem]
tavily-python
